Fix VPATH builds
The dist tarball should now include the share/ directory, which it was
omitting previously.  For whatever reason, I had to remove the $(RM) and
$(RMDIR) macros, in favor of the POSIX names rm and rmdir.  Seems to
work on my Linux envs, but may cause problems down the road.

Also, fixed a missing srcdir for VPATH builds, which fixes #11 - `make
distcheck` now works.