resolve drivepath in linux collector (#21)
The Linux collector sets KEYS to map partitions to drivers. IE; we collect
metrics on partition usage and space allocation via the mount point, but
all IO statistics are collected via the drive path.

When a Linux system is running device-mapper which is usually the case when
LVM or multipath, mounts are usually mounted via their device-mapper name and
not the block device name in /dev. Thus the KEY generated in the partition
space usage snippet coordinates back to the device mapper name, and not the
block device name, IO-stats are stored for block device, so we are mapping
to a namespace without data.

This patch resolves the device mapper name back to the linux block device
and drops the device-mapper key in favor of the block device.